+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Adds Doorkeeper Generator with Devise. ([@TheZero0-ctrl][])
 * Adds Avo generator. ([@mausamp][])
 * Adds Sentry generator. ([@mausamp][])
+* Adds Dotenv generator. ([@mausamp][])
 
 ## 0.13.0 (March 26th, 2024)
 * Adds Letter Opener generator. ([@coolprobn][])